+++ src/Models/CertificateManager.php
@@ -2,29 +2,29 @@
 
 namespace Cyberfusion\ClusterApi\Models;
 
+use Cyberfusion\ClusterApi\Enums\ProviderNames;
 use Cyberfusion\ClusterApi\Support\Arr;
 use Cyberfusion\ClusterApi\Contracts\Model;
 use Cyberfusion\ClusterApi\Support\Validator;
 
-class Certificate extends ClusterModel implements Model
+class CertificateManager extends ClusterModel implements Model
 {
-    private string $mainCommonName = '';
+    private ?string $mainCommonName = '';
     private array $commonNames = [];
-    private string $certificate;
-    private string $caChain;
-    private string $privateKey;
-    private string $expiresAt;
+    private string $providerName;
+    private ?string $requestCallbackUrl = null;
+    private ?int $certificateId = null;
     private ?int $id = null;
     private ?int $clusterId = null;
     private ?string $createdAt = null;
     private ?string $updatedAt = null;
 
-    public function getMainCommonName(): string
+    public function getMainCommonName(): ?string
     {
         return $this->mainCommonName;
     }
 
-    public function setMainCommonName(string $mainCommonName): Certificate
+    public function setMainCommonName(?string $mainCommonName): CertificateManager
     {
         $this->mainCommonName = $mainCommonName;
 
@@ -36,7 +36,7 @@ class Certificate extends ClusterModel implements Model
         return $this->commonNames;
     }
 
-    public function setCommonNames(array $commonNames): Certificate
+    public function setCommonNames(array $commonNames): CertificateManager
     {
         Validator::value($commonNames)
             ->unique()
@@ -47,65 +47,42 @@ class Certificate extends ClusterModel implements Model
         return $this;
     }
 
-    public function getCertificate(): string
+    public function getProviderName(): string
     {
-        return $this->certificate;
+        return $this->providerName;
     }
 
-    public function setCertificate(string $certificate): Certificate
+    public function setProviderName(string $providerName): CertificateManager
     {
-        Validator::value($certificate)
-            ->maxLength(65535)
-            ->pattern('^[a-zA-Z0-9-_\+\/=\n ]+$')
+        Validator::value($providerName)
+            ->valueIn(ProviderNames::AVAILABLE)
             ->validate();
 
-        $this->certificate = $certificate;
+        $this->providerName = $providerName;
 
         return $this;
     }
 
-    public function getCaChain(): string
+    public function getRequestCallbackUrl(): ?string
     {
-        return $this->caChain;
+        return $this->requestCallbackUrl;
     }
 
-    public function setCaChain(string $caChain): Certificate
+    public function setRequestCallbackUrl(?string $requestCallbackUrl): CertificateManager
     {
-        Validator::value($caChain)
-            ->maxLength(65535)
-            ->pattern('^[a-zA-Z0-9-_\+\/=\n ]+$')
-            ->validate();
-
-        $this->caChain = $caChain;
+        $this->requestCallbackUrl = $requestCallbackUrl;
 
         return $this;
     }
 
-    public function getPrivateKey(): string
+    public function getCertificateId(): ?int
     {
-        return $this->privateKey;
+        return $this->certificateId;
     }
 
-    public function setPrivateKey(string $privateKey): Certificate
+    public function setCertificateId(?int $certificateId): CertificateManager
     {
-        Validator::value($privateKey)
-            ->maxLength(65535)
-            ->pattern('^[a-zA-Z0-9-_\+\/=\n ]+$')
-            ->validate();
-
-        $this->privateKey = $privateKey;
-
-        return $this;
-    }
-
-    public function getExpiresAt(): string
-    {
-        return $this->expiresAt;
-    }
-
-    public function setExpiresAt(string $expiresAt): Certificate
-    {
-        $this->expiresAt = $expiresAt;
+        $this->certificateId = $certificateId;
 
         return $this;
     }
@@ -115,7 +92,7 @@ class Certificate extends ClusterModel implements Model
         return $this->id;
     }
 
-    public function setId(?int $id): Certificate
+    public function setId(?int $id): CertificateManager
     {
         $this->id = $id;
 
@@ -127,7 +104,7 @@ class Certificate extends ClusterModel implements Model
         return $this->clusterId;
     }
 
-    public function setClusterId(?int $clusterId): Certificate
+    public function setClusterId(?int $clusterId): CertificateManager
     {
         $this->clusterId = $clusterId;
 
@@ -139,7 +116,7 @@ class Certificate extends ClusterModel implements Model
         return $this->createdAt;
     }
 
-    public function setCreatedAt(?string $createdAt): Certificate
+    public function setCreatedAt(?string $createdAt): CertificateManager
     {
         $this->createdAt = $createdAt;
 
@@ -151,22 +128,21 @@ class Certificate extends ClusterModel implements Model
         return $this->updatedAt;
     }
 
-    public function setUpdatedAt(?string $updatedAt): Certificate
+    public function setUpdatedAt(?string $updatedAt): CertificateManager
     {
         $this->updatedAt = $updatedAt;
 
         return $this;
     }
 
-    public function fromArray(array $data): Certificate
+    public function fromArray(array $data): CertificateManager
     {
         return $this
             ->setMainCommonName(Arr::get($data, 'main_common_name', ''))
             ->setCommonNames(Arr::get($data, 'common_names', []))
-            ->setCertificate(Arr::get($data, 'certificate'))
-            ->setCaChain(Arr::get($data, 'ca_chain'))
-            ->setPrivateKey(Arr::get($data, 'private_key'))
-            ->setExpiresAt(Arr::get($data, 'expires_at'))
+            ->setProviderName(Arr::get($data, 'provider_name'))
+            ->setRequestCallbackUrl(Arr::get($data, 'request_callback_url'))
+            ->setCertificateId(Arr::get($data, 'certificate_id'))
             ->setId(Arr::get($data, 'id'))
             ->setClusterId(Arr::get($data, 'cluster_id'))
             ->setCreatedAt(Arr::get($data, 'created_at'))
@@ -178,10 +154,9 @@ class Certificate extends ClusterModel implements Model
         return [
             'main_common_name' => $this->getMainCommonName(),
             'common_names' => $this->getCommonNames(),
-            'certificate' => $this->getCertificate(),
-            'ca_chain' => $this->getCaChain(),
-            'private_key' => $this->getPrivateKey(),
-            'expires_at' => $this->getExpiresAt(),
+            'provider_name' => $this->getProviderName(),
+            'request_callback_url' => $this->getRequestCallbackUrl(),
+            'certificate_id' => $this->getCertificateId(),
             'id' => $this->getId(),
             'cluster_id' => $this->getClusterId(),
             'created_at' => $this->getCreatedAt(),